[Feat]: DrawingPanel UI와 배치용 객체로 분리
