+++ ItsMyFirstLoveProject/Assets/01.Scenes/Test/KJS/DrawingPanel.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1b7096b0f09612a488ca0852b4a8c39e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
