+++ ItsMyFirstLoveProject/Assets/01.Scenes/Test/KJS/DrawingPanel/DrawingPanelTest.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9929608f4a9306a4bb4bc4cecb2f2683
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
