+++ ItsMyFirstLoveProject/Assets/02.Scripts/DrawingPanel.cs
@@ -1,175 +1,29 @@
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.EventSystems;
 using UnityEngine.UI;
 using GameFile;
 
 public class DrawingPanel : MonoBehaviour
 {
-    private enum Color
-    {
-        BLACK,
-        WHITE,
-        PINK,
-        SKYBLUE
-    }
-
-    [SerializeField] private List<Material> _lineMaterials = new List<Material>(); 
     private Stack<GameObject> _lineStack = new Stack<GameObject>();
     private RawImage _image;
-//--------------------- For Test --------------------------
-    // 펜 색상 선택
-    [SerializeField] private int _selectColor;
-//---------------------------------------------------------
-
-    // For RayCast
-    private LineRenderer curLine;
-    private int positionCount = 2;
-    private Vector3 PrevPos = Vector3.zero;
-    private RaycastHit hit;
-    private bool _canvasTouch;
 
     private void Start() 
     {
         _image = GetComponentInChildren<RawImage>();
     }
-
-    private void Update()
+    
+    private void OnEnable() 
     {
-        Drawtouch();
-    }
-
-    /// <summary>
-    /// 메인카메라 기준 이미지 캡쳐 후 텍스쳐 로드
-    /// </summary>
-    public void SaveImage()
-    {
-        Img.Save(Camera.main.gameObject, Img.Size.Draw, Img.Name.Draw);
-        EraserAll();
         LoadImage();
     }
 
-    /// <summary>
-    /// 가장 최근에 저장했던 이미지 텍스쳐 로드
-    /// </summary>
-    public void LoadImage()
+    // 가장 최근에 저장했던 이미지 텍스쳐 로드
+    private void LoadImage()
     {
         _image.texture = Img.TextureLoad();
     }
 
-    // 캔버스 건물 배치 (미완. 추후 배치 관련 기획안 수립시 작업요망)
-    private void PositionSetToBuilding(GameObject Building)
-    {
-        // TODO : 건물에 배치하는 위치나 방식에 따라 추가 구현해야 함.
-        transform.parent = Building.transform;
-    }
 
-    /// <summary>
-    /// 선 전체 지우기
-    /// </summary>
-    public void EraserAll()
-    {
-        for(int i = 0; i < _lineStack.Count; i++)
-        {
-            Eraser();
-        }
-    }
-
-    /// <summary>
-    /// 지우기 기능(실행 취소)
-    /// </summary>
-    public void Eraser()
-    {
-        Destroy(_lineStack.Pop());
-    }
-
-    // 터치 동작시 Canvas에 선 그리는 함수
-    private void Drawtouch()
-    {
-        if (Input.touchCount > 0)
-        {
-            Touch touch = Input.GetTouch(0);
-            Ray ray = Camera.main.ScreenPointToRay(touch.position);
-
-            if(Physics.Raycast(ray, out hit))
-            {
-                if(hit.collider.name == "DrawCanvas")
-                {
-                    Vector3 touchPos = Camera.main.ScreenToWorldPoint(new Vector3(touch.position.x, touch.position.y, hit.collider.transform.position.z - 0.1f));
-                    if(touch.phase == TouchPhase.Began && _canvasTouch == false)
-                    {
-                        createLine(touchPos);
-                        _canvasTouch = true;
-                    }
-                    if(touch.phase == TouchPhase.Moved && _canvasTouch == true)
-                    {
-                        connectLine(touchPos);
-                    }
-                    
-                }
-                if(touch.phase == TouchPhase.Ended)
-                {
-                    _canvasTouch = false;
-                }
-            }
-        }
-    }
-
-    private void createLine(Vector3 touchPos)
-    {
-        positionCount = 2;  
-        GameObject line = new GameObject("Line");
-        _lineStack.Push(line);
-        LineRenderer lineRend = line.AddComponent<LineRenderer>();
-
-        line.transform.parent = transform;
-        line.transform.position = touchPos;
-
-        lineRend.startWidth = 0.01f;
-        lineRend.endWidth = 0.01f;
-        lineRend.numCornerVertices = 5;
-        lineRend.numCapVertices = 5;
-        lineRend.material = _lineMaterials[_selectColor];
-        lineRend.SetPosition(0, touchPos);
-        lineRend.SetPosition(1, touchPos);
-
-        curLine = lineRend;
-    }
-
-    private void connectLine(Vector3 touchPos)
-    {
-        if(PrevPos != null && Mathf.Abs(Vector3.Distance(PrevPos, touchPos)) >= 0.001f)
-        {
-            PrevPos = touchPos;
-            positionCount++;
-            curLine.positionCount = positionCount;
-            curLine.SetPosition(positionCount - 1, touchPos);
-        }
-    }
-
-
-// 미사용 메서드. 추후 확장 기능을 위해 임시 보관. 미사용시 삭제해도 무방함------------------------------------
-
-    // 터치(레이캐스트) 획 지우기 작성중이었던 코드. 실행 취소로 대체함.
-    
-    // private void Eraser()
-    // {
-    //     if (Input.touchCount > 0)
-    //     {
-    //         Touch touch = Input.GetTouch(0);
-    //         Ray ray = Camera.main.ScreenPointToRay(touch.position);
-    //         Vector3 touchPos = Camera.main.ScreenToWorldPoint(new Vector3(touch.position.x, touch.position.y, hit.collider.transform.position.z - 0.1f));
-
-    //         Debug.Log(touchPos);
-
-    //         if(Physics.Raycast(ray, out hit, 100))
-    //         {
-    //             if(hit.collider.name == "Line")
-    //             {
-    //                 Destroy(hit.collider.gameObject);
-    //             }
-    //         }
-    //     }
-    // }
 
 }
