+++ ItsMyFirstLoveProject/Assets/02.Scripts/DrawingPanelUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 694dfc6474366a14289fd1fcbe69c33b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
