+++ ItsMyFirstLoveProject/Assets/03.Prefabs/DrawingPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5071e047eeae09a4d9ba0ca1892d238d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
