+++ ItsMyFirstLoveProject/Assets/03.Prefabs/UI/InGame/DrawingPanelUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0a0267dd3222b084ca4347ae6130311b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
