+++ ItsMyFirstLoveProject/Assets/05.Materials/DrawingPanel.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de8b0d88d32552f409b11fb74f03a357
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
