+++ ItsMyFirstLoveProject/Assets/05.Materials/DrawingPanel/Drawing_Black.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3bc2d31e586a6a741b3bc5b59a846ac4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
