+++ ItsMyFirstLoveProject/Assets/05.Materials/DrawingPanel/Drawing_Pink.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a4f5ee0d2c225342b8911afa01c6fe1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
