+++ ItsMyFirstLoveProject/Assets/05.Materials/DrawingPanel/Drawing_SkyBlue.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 654906c024434c1499858980442b5964
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
