+++ ItsMyFirstLoveProject/Assets/05.Materials/DrawingPanel/Drawing_White.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07f18b6e164459c4c907761b33b87beb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
